Replace usage of StatusScreen#addComponentOf(String) and its using Component.text().
Add component creating methods to TownyFormatter to generate Components
representing basic key and key-value colouring options.

Make use of TownyFormatter#initialize, to parse the colouring options on
load, out of legacy chat slugs or using hex or if needed a fallback.

Issues:
  - Lots of new lines are being made.
  - Hex is still not supported when a string is parsed via
Translation#of and goes through Colors#translateColourCodes, which is
then improperly coloured when it is run through Component#text.
     - This will require going through most strings to make sure they're
not coloured, or stripping them of their colour.